Documents/GitHub/creative-intel-app/requirements.txt
# API and Web Framework
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1

# Machine Learning and Computer Vision
torch==2.2.0
torchvision==0.17.0
transformers==4.37.2
sentence-transformers==2.5.1
opencv-python==4.9.0.80
pytesseract==0.3.10
Pillow==10.2.0

# Vector Database
pinecone-client==3.0.2

# OpenAI
openai==1.12.0

# Utilities
numpy==1.26.4
pandas==2.2.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
httpx==0.26.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0 